Rename simpleCommands to examplelessCommands.
