+++ src/settings.ts
@@ -50,14 +50,14 @@ export default class SettingTab extends PluginSettingTab {
             })
 
         new Setting(containerEl)
-            .setName(t.settings.simpleCommands.title)
-            .setDesc(t.settings.simpleCommands.description)
+            .setName(t.settings.examplelessCommands.title)
+            .setDesc(t.settings.examplelessCommands.description)
             .addToggle(toggle => {
                 toggle
-                    .setTooltip(t.settings.simpleCommands.tooltip)
-                    .setValue(this.plugin.settings.simpleCommands)
+                    .setTooltip(t.settings.examplelessCommands.tooltip)
+                    .setValue(this.plugin.settings.examplelessCommands)
                     .onChange(async value => {
-                        this.plugin.settings.simpleCommands = value
+                        this.plugin.settings.examplelessCommands = value
                         await this.plugin.saveSettings()
                     })
             })
